+++ exporter/opencensusexporter/factory.go
@@ -52,7 +52,7 @@ func createDefaultConfig() configmodels.Exporter {
 	}
 }
 
-func createTraceExporter(ctx context.Context, params component.ExporterCreateParams, config configmodels.Exporter) (component.TraceExporter, error) {
+func createTraceExporter(ctx context.Context, params component.ExporterCreateParams, config configmodels.Exporter) (component.TracesExporter, error) {
 	oCfg := config.(*Config)
 	return newTraceExporter(ctx, oCfg, params.Logger)
 }