+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Rename consumer.TraceConsumer to consumer.TracesConsumer #1974
 - Rename component.TraceProcessor to component.TracesProcessor #1976
+- Rename component.TraceExporter to component.TracesExporter #1975
 
 ## v0.13.0 Beta